chore: cleanup and move function to correct adapter
