+++ src/services/lease-service/adapters/listing-adapter.ts
@@ -132,11 +132,22 @@ const getApplicantsByContactCodeAndRentalObjectCode = async (contactCode: string
     .first();
 }
 
+const applicationExists = async (contactCode: string, listingId: number) => {
+  const result = await db('applicant')
+    .where({
+      ContactCode: contactCode,
+      ListingId: listingId
+    })
+    .first();
+  return !!result; // Convert result to boolean: true if exists, false if not
+};
+
 export {
   createListing,
   createApplication,
   getListingByRentalObjectCode,
   getAllListingsWithApplicants,
   getApplicantsByContactCode,
-  getApplicantsByContactCodeAndRentalObjectCode
+  getApplicantsByContactCodeAndRentalObjectCode,
+  applicationExists
 }
