+++ src/services/lease-service/adapters/xpand/tenant-lease-adapter.ts
@@ -338,95 +338,6 @@ const isLeaseActive = (lease: Lease | PartialLease): boolean => {
   )
 }
 
-const createListing = async (listingData: Listing) => {
-  await db('Listing').insert({
-    RentalObjectCode: listingData.rentalObjectCode,
-    Address: listingData.address,
-    DistrictCaption: listingData.districtCaption,
-    DistrictCode: listingData.districtCode,
-    BlockCaption: listingData.blockCaption,
-    BlockCode: listingData.blockCode,
-    MonthlyRent: listingData.monthlyRent,
-    ObjectTypeCaption: listingData.objectTypeCaption,
-    ObjectTypeCode: listingData.objectTypeCode,
-    RentalObjectTypeCaption: listingData.rentalObjectTypeCaption,
-    RentalObjectTypeCode: listingData.rentalObjectTypeCode,
-    PublishedFrom: listingData.publishedFrom,
-    PublishedTo: listingData.publishedTo,
-    VacantFrom: listingData.vacantFrom,
-    Status: listingData.status,
-    WaitingListType: listingData.waitingListType,
-  });
-}
-
-/**
- * Checks if a listing already exists based on unique criteria.
- * 
- * @param {string} rentalObjectCode - The rental object code of the listing (originally from xpand)
- * @returns {Promise<Listing>} - Promise that resolves to the existing listing if it exists.
- */
-const getListingByRentalObjectCode = async (rentalObjectCode: string): Promise<Listing> => {
-  const existingListing = await db('Listing')
-    .where({
-      RentalObjectCode: rentalObjectCode
-    })
-    .first();
-
-  return existingListing;
-};
-
-const createApplication = async (applicationData: Applicant) => {
-  console.log(applicationData);
-  await db('applicant').insert({
-    Name: applicationData.name,
-    ContactCode: applicationData.contactCode,
-    ApplicationDate: applicationData.applicationDate,
-    ApplicationType: applicationData.applicationType,
-    Status: applicationData.status,
-    ListingId: applicationData.listingId,
-  });
-}
-
-const getAllListingsWithApplicants = async () => {
-  const listings = await db('Listing').select('*');
-
-  for (let listing of listings) {
-    const applicants = await db('Applicant')
-      .where('listingId', listing.Id)
-      .select('*');
-
-    listing.applicants = applicants;
-  }
-
-  return listings;
-};
-
-const getApplicantsByContactCode = async (contactCode: string) => {
-  return db('Applicant')
-    .where({ ContactCode: contactCode });
-}
-
-const getApplicantsByContactCodeAndRentalObjectCode = async (contactCode: string, rentalObjectCode: string) => {
-  return db('Applicant')
-    .where({
-      ContactCode: contactCode,
-      RentalObjectCode: rentalObjectCode
-    })
-    .first();
-}
-
-const applicationExists = async (contactCode: string, listingId: number) => {
-  const result = await db('applicant')
-    .where({
-      ContactCode: contactCode,
-      ListingId: listingId
-    })
-    .first();
-  return !!result; // Convert result to boolean: true if exists, false if not
-};
-
-
-
 export {
   getLease,
   getLeases,
@@ -434,11 +345,5 @@ export {
   getLeasesForNationalRegistrationNumber,
   getContactByNationalRegistrationNumber,
   getContactByContactCode,
-  isLeaseActive,
-  createListing,
-  createApplication,
-  getListingByRentalObjectCode,
-  getAllListingsWithApplicants,
-  getApplicantsByContactCode,
-  getApplicantsByContactCodeAndRentalObjectCode
+  isLeaseActive
 }
